+++ generators/app/tasks/appSetup/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable max-statements */
 const ora = require('ora');
 
 const packageJsonScripts = require('./packageJsonScripts');
@@ -5,6 +6,7 @@ const eslintSetup = require('./eslintSetup');
 const baseFilesTemplate = require('./baseFilesTemplate');
 const iosAppIcons = require('./iosAppIcons');
 const androidProjectSetup = require('./androidProjectSetup');
+const createDotEnvFilesLocally = require('./createDotEnvFilesLocally');
 const iosProjectSetup = require('./iosProjectSetup');
 const disableLandscapeOrientation = require('./disableLandscapeOrientation');
 const pushNotificationsFeatureFiles = require('./pushNotificationsFeatureFiles');
@@ -16,7 +18,7 @@ const loginFeatureFiles = require('./loginFeatureFiles');
 const bitriseFeatureFiles = require('./bitriseFeatureFiles');
 const enableFullscreen = require('./tabletSetup');
 const babelConfigSetup = require('./babelConfigSetup');
-const editBundleIdentifier = require('./editBundleIdentifier');
+const cleanTargetsFromPods = require('./cleanTargetsFromPods');
 const prettierrcConfigSetup = require('./prettierrcConfigSetup');
 const splashScreenSetup = require('./splashScreenSetup');
 const firebasePerformanceSetup = require('./firebasePerformanceSetup');
@@ -26,6 +28,9 @@ module.exports = function index() {
     spinner: 'bouncingBall',
     text: 'Creating project boilerplate'
   }).start();
+
+  // ----------------     add envs files     ----------------
+  createDotEnvFilesLocally.bind(this)();
 
   // ----------------     add package.json scripts     ----------------
   packageJsonScripts.bind(this)();
@@ -45,14 +50,12 @@ module.exports = function index() {
   // ----------------     prettierrc setup     ----------------
   prettierrcConfigSetup.bind(this)();
 
-  // ----------------     add react-native-config to app/build.gradle     ----------------
-  // ----------------     add react-native-screen to app/build.gradle     ----------------
-  // ----------------     add react-native-gesture-handler to MainActivity    ----------------
+  // ----------------     add Android project configurationn    ----------------
   androidProjectSetup.bind(this)();
 
-  // ----------------     fix bundle identifier     ----------------
-  editBundleIdentifier.bind(this)();
+  // ----------------     add iOS project configuration and clean unused targets    ----------------
   iosProjectSetup.bind(this)();
+  cleanTargetsFromPods.bind(this)();
 
   // ----------------     disable landscape orientiation for both android and ios     ----------------
   if (this.features.landscape) {
@@ -67,9 +70,11 @@ module.exports = function index() {
     loginFeatureFiles.bind(this)();
   }
 
+  // ----------------     Features: Bitrise    ----------------
   if (this.features.bitrise) {
     bitriseFeatureFiles.bind(this)();
   }
+
   // ----------------     Features: Firebase    ----------------
   if (
     this.features.crashlytics ||