+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.8.0] - 2020-04-XX
+
+- Added multiple env configuration (Debug, QA, Stage and Production) for Android and iOS projects
+- Removed unused targets from iOS project and Podfile (ruby script)
+- Removed temporary fix for the `update_bundle_identifier` Fastlane lane
+
 ## [0.7.3] - 2020-04-03
 
 - Fixed Push Notifications config on `AndroidManifest`